// src/generated/companionImageFocalManifest.ts
export interface BundledCompanionImageFocalManifestEntry {
  focalX: number;
  focalY: number;
  width: number;
  height: number;
}